# Cache some properties on Dataset