--- a/h5py/_hl/dataset.py
+++ b/h5py/_hl/dataset.py
@@ -278,10 +278,20 @@ class Dataset(HLObject):
         return self.id.rank
 
     @property
-    @with_phil
     def shape(self):
         """Numpy-style shape tuple giving dataset dimensions"""
-        return self.id.shape
+        if 'shape' in self._cache_props:
+            return self._cache_props['shape']
+
+        with phil:
+            shape = self.id.shape
+
+        # If the file is read-only, cache the shape to speed-up future uses.
+        # This cache is invalidated by .refresh() when using SWMR.
+        if self._readonly:
+            self._cache_props['shape'] = shape
+        return shape
+
     @shape.setter
     @with_phil
     def shape(self, shape):
@@ -289,12 +299,21 @@ class Dataset(HLObject):
         self.resize(shape)
 
     @property
-    @with_phil
     def size(self):
         """Numpy-style attribute giving the total dataset size"""
-        if is_empty_dataspace(self.id):
-            return None
-        return numpy.prod(self.shape, dtype=numpy.intp)
+        if 'size' in self._cache_props:
+            return self._cache_props['size']
+
+        if self._is_empty:
+            size = None
+        else:
+            size = numpy.prod(self.shape, dtype=numpy.intp)
+
+        # If the file is read-only, cache the size to speed-up future uses.
+        # This cache is invalidated by .refresh() when using SWMR.
+        if self._readonly:
+            self._cache_props['size'] = size
+        return size
 
     @property
     @with_phil
@@ -395,7 +414,7 @@ class Dataset(HLObject):
         return self._extent_type == h5s.NULL
 
     @with_phil
-    def __init__(self, bind):
+    def __init__(self, bind, readonly=False):
         """ Create a new Dataset object by binding to a low-level DatasetID.
         """
         if not isinstance(bind, h5d.DatasetID):
@@ -405,6 +424,8 @@ class Dataset(HLObject):
         self._dcpl = self.id.get_create_plist()
         self._dxpl = h5p.create(h5p.DATASET_XFER)
         self._filters = filters.get_filters(self._dcpl)
+        self._readonly = readonly
+        self._cache_props = {}
         self._local = local()
         self._local.astype = None
 
@@ -532,7 +553,7 @@ class Dataset(HLObject):
 
         # === Check for zero-sized datasets =====
 
-        if numpy.product(self.shape, dtype=numpy.ulonglong) == 0:
+        if self.size == 0:
             # Check 'is Ellipsis' to avoid equality comparison with an array:
             # array equality returns an array, not a boolean.
             if args == () or (len(args) == 1 and args[0] is Ellipsis):
@@ -817,6 +838,7 @@ class Dataset(HLObject):
             library version >=1.9.178
             """
             self._id.refresh()
+            self._cache_props.clear()
 
     if hasattr(h5d.DatasetID, "flush"):
         @with_phil