In #1391, we noticed that repeated small reads from the same dataset are surprisingly slow. Profiling showed that quite a bit of time was wasted re-fetching and calculating things which don't change from read to read.

First, the headline figures:

- Simple `ds[0:10]` reads are about 2x faster (1.2 s -> 0.55 s)
- `read_direct()` about 25% faster (0.98 s -> 0.75 s)

Times are for 10k reads of 10 numbers each from a 1D integer dataset. [Details in this comment](https://github.com/h5py/h5py/issues/1391#issuecomment-542080099), but the timings given there are done with a different method.

The fastest I've got with Python code calling the low-level API directly is 0.22 s.

-----

For properties which can't change (like whether a dataset is empty), I've used a standard [@cached_property](https://github.com/pydanny/cached-property) decorator. They'll be calculated once on first access, then saved.

Dataset shape and size can change, either if the file is writeable (through `H5Dset_extent`) or if it's being read in SWMR mode (when `H5Drefresh` is called). I've taken a compromise position: 

* For SWMR mode the `ds.refresh()` call also clears the cache. This assumes that people keeping a `Dataset` object around will use its refresh method, not the low-level API.
* If the file is writeable, it doesn't use the cache at all, so it can't get outdated if you modify it with the low-level API.

The code could be simpler if we cached shape & size for writeable files as well - but it would mean that after manipulating a dataset with the low-level API you need to call `.refresh()` or get a new high-level Dataset object.