@takluyver 2019-10-15
Another tweak: using a simple Python function in place of `numpy.product()` in `SimpleSelection.broadcast()`.

- `ds[0:10]` reading unchanged
- `read_direct()` gets about 30% faster again (0.75 s -> 0.52 s). This is almost a 2x speedup from before this PR.

-----

With small amounts of data, `numpy.product()` is about an order of magnitude slower than the hand-coded `product()` function I added. E.g. with a tuple of 2 numbers, numpy takes ~8 µs, and the custom function about 200 ns.

Of course, this only works for small data, and possibly only for data that's not already in an array. But in various places we have one number per dimension, like this, and [HDF5 only allows up to 32 dimensions](https://support.hdfgroup.org/HDF5/faq/limits.html). Even if that limit were removed, I suspect there are very few use cases which need more dimensions.