@takluyver 2019-10-23
I'll merge this in another day or two unless anyone objects or wants more time to review it.