@takluyver 2019-10-25
Merging on the strength of @tacaswell's approval and lack of objections. There's still time to make changes before release.